+++ lib/FieldDefinitions/Email.js
@@ -14,6 +14,8 @@ var _react2 = _interopRequireDefault(_react);
 
 var _immutable = require('immutable');
 
+var _utils = require('../utils');
+
 var _reactDnd = require('react-dnd');
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
@@ -101,54 +103,8 @@ var Email = function (_Component) {
       var currentValue = formValues.get(config.name, '');
       config = _extends({ currentValue: currentValue }, config);
       handleCascadeKeywordClick(config);
-    }, _this.emailValidation = function (email) {
-      // If email is null, undefined, empty array or string
-      if (!email) {
-        return false;
-      }
-      email = String(email).trim();
-
-      // https://www.mailboxvalidator.com/resources/articles/acceptable-email-address-syntax-rfc/
-      var emailInput = email.split('@');
-      var emailLocal = emailInput[0].trim();
-
-      // email does not contain "@"
-      if (emailInput.length === 1) {
-        return false;
-      }
-
-      // validating local part of the email
-      if (emailLocal.length > 64 || emailLocal.length < 1 || emailLocal.includes('..') || emailLocal[0] === '.' || emailLocal[emailLocal.length - 1] === '.') {
-        return false;
-      }
-
-      var domainPart = emailInput[1].trim();
-      var emailDomains = domainPart.split('.');
-
-      // Entire domain part of the email does not contain "." or is greater than 255 characters
-      if (emailDomains.length < 2 || domainPart.length > 255) {
-        return false;
-      }
-
-      var validateDomainSegment = function validateDomainSegment(segment) {
-        // A true here means the domain segment failed validation
-        return segment.length > 63 || segment.length < 1 || segment[0] === '-' || segment[segment.length - 1] === '-';
-      };
-
-      // emailDomains MAY consist of subdomain1.subDomain2.domain.topLevelDomain need to validate each domain segment
-      if (emailDomains.some(validateDomainSegment)) {
-        return false;
-      }
-
-      // Top level domains may not be entirely numeric
-      var tld = emailDomains[emailDomains.length - 1];
-      if (tld.length < 2 || tld.match(/^[0-9]+$/)) {
-        return false;
-      }
-
-      return true;
     }, _this.generateValidationError = function (value) {
-      if (value) return !_this.emailValidation(value) && _this.state.displayError && '* Invalid email';
+      if (value) return !(0, _utils.emailValidator)(value) && _this.state.displayError && '* Invalid email';
     }, _this.handleOnChange = function () {}, _this.render = function () {
       var _this$props5 = _this.props,
           inline = _this$props5.inline,
@@ -256,7 +212,7 @@ var Email = function (_Component) {
           fontSize: '9pt',
           marginLeft: 3,
           marginTop: -1,
-          color: !_this.emailValidation(value) && '#ec1c24',
+          color: !(0, _utils.emailValidator)(value) && '#ec1c24',
           marginRight: 5
         }
       };
@@ -317,9 +273,6 @@ var Email = function (_Component) {
 
     // This seems like an antipattern - a function and a prop named the same thing with the function
     // calling the prop provided function, maybe the component function should be called onCascadeKeywordClick???
-
-
-    // This could potentially be extracted into a helper function for reuse across projects.
 
   }]);
