+++ lib/GridFormBuilder.js
@@ -126,6 +126,8 @@ var _Percentage = require('./FieldDefinitions/Percentage');
 
 var _Percentage2 = _interopRequireDefault(_Percentage);
 
+var _utils = require('./utils');
+
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
 function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
@@ -307,13 +309,21 @@ var FormBuilder = function (_Component) {
             required = _config$required2 === undefined ? false : _config$required2,
             name = config.name,
             _config$label = config.label,
-            label = _config$label === undefined ? name : _config$label;
+            label = _config$label === undefined ? name : _config$label,
+            type = config.type;
 
         if (required && formValues.get(name, '').length === 0) {
           reasons.push({
             reason: 'required',
             message: label + ' cannot be blank.',
             description: 'The field ' + name + ' is marked as required, but its value is empty.'
+          });
+        }
+        if (type === 'email' && !(0, _utils.emailValidator)(formValues.get(name, ''))) {
+          reasons.push({
+            reason: 'incorrect format',
+            message: label + ' is invalid',
+            description: 'The field ' + name + ' has an invalid email'
           });
         }
       });
